Add support for float values

+++ python/src/config.py
@@ -2,7 +2,7 @@ class Config:
     def __init__(self, cParams: dict):
         self.filePath: str = cParams.get("file", None)
         self.dirPath: str = cParams.get("dir", None)
-        self.borderAmount: int = cParams.get("border", 5)
+        self.borderAmount: float = cParams.get("border", 5)
         self.colour: str = cParams.get("colour", "white")
         self.useLong: bool = cParams.get("useLong", False)
